platform/chrome: cros_ec_proto: Also return invalid command in cros_ec_cmd_xfer_status
The commit 9798ac6d32c1 ("mfd: cros_ec: Add cros_ec_cmd_xfer_status()
helper") was introduced to remove duplicated code in several users of
the cros_ec_cmd_xfer function. The idea behind that commit is mask all
the EC errors as protocol error when the EC command fails, what is
really useful because as a user you probably don't care of the exact
error reported by the EC, you only want to know if the command succeeded
or not.

That works for most of the errors returned by the EC, but we might be
also interested on report the EC_RES_INVALID_COMMAND, which happens when
the command is not supported by the EC. An use case for this is when you
need to show/hide a sysfs attribute for a specific command and you don't
have a EC feature flag that tells you that the command is supported or
not.

While here, also document properly the cros_ec_cmd_xfer_status
function.

Signed-off-by: Enric Balletbo i Serra <enric.balletbo@collabora.com>
Reviewed-by: Guenter Roeck <groeck@chromium.org>