CHROMIUM: cros_ec_istp loops until events are gone
The ISH FW supports a flags that allows the kernel to continue to
query for more MKBP events until they are all consumed. This is the
preferred method for consuming events after receiving a MKBP event.

Also record the timestamp as close to the hardware event as possible
without breaking encapsulation boundaries.

BRANCH=none
BUG=b:129159505
TEST=AP is honoring flag and calling to get more events if needed

Change-Id: Id8f433bc08ae5329502668b443da96eb26a8b099
Signed-off-by: Jett Rink <jettrink@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1555219
Reviewed-by: Enrico Granata <egranata@chromium.org>
Reviewed-by: Mathew King <mathewk@chromium.org>

[rebase54(gwendal): File location changed]