CHROMIUM: ARM: rockchip: pm: Support deep suspend
The mainline Linux kernel has some placeholder code in "pm.c" for
supporting deep suspend but it's not actually selectable because
ROCKCHIP_ARM_OFF_LOGIC_NORMAL is always passed to rk3288_slp_mode_set.
...it also doesn't work due to the lack of any code to re-init SDRAM.

This patch actually makes deep suspend work and also makes it the
default, introducing a module parameter to go back to the non-deep
suspend mode (which you'd only want to do if for some reason deep
suspend wasn't working for you).

This code is mostly copied from the downstream Chrome OS 3.14 tree
with the following changes:

* The DDR Frequency changing code is currently stripped out of the
  "embedded" directory.  It'll be re-introduced in a later patch.
* In a previous change I changed the "rockchip,rk3288-noc" name to
  "rockchip,rk3288-service-bus" to match the TRM.  This code adapts to
  that.
* Upstream changed virt_to_phys() to __pa_symbol() in "pm.c" so I did
  too.

In general this code is not considered upstreamable (and thus this
patch is marked CHROMIUM) due to upstream not liking the way we do
embedded blobs [1].  Some other attempts have been made to come up
with better ways to do embedded blobs [2] but I know none that have
actually succeeded and landed their code.  By now people seem to have
given up since on newer systems the answer is to just jam all this
type of code into the firmware and hide it from the NAKing eyes of
kernel maintainers.

[1] https://lkml.kernel.org/r/1417461694-5129-1-git-send-email-dianders@chromium.org
[2] https://lkml.kernel.org/r/1459668234-16033-2-git-send-email-alexandre.belloni@free-electrons.com

BUG=chromium:941638
TEST=suspend_stress_test

Change-Id: I5c179ff9392a37218e27d6162a041c69f14026ce
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1570555
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>
[rebase54(groeck):
	sleep.S modified upstream and locally removed
	arch_counter_get_cntpct -> __arch_counter_get_cntpct
 squashed commits:
	FIXUP: CHROMIUM: ARM: rockchip: pm: Support deep suspend
	FIXUP: CHROMIUM: ARM: rockchip: pm: Support deep suspend (take 2)
	]
Signed-off-by: Guenter Roeck <groeck@chromium.org>