parisc: Remove lock code to serialize TLB operations in pacache.S
TLB operations only need to be serialized on machines with the Merced (Stretch)
bus. The only machines in this category are L and N class, and they require a
64-bit PA 2.0 kernel. On these machines, we use local TLB purges in the
tmpalias routines.
We don't need to serialize TLB purges on all other machines. Thus, the
lock/unlock code can be removed when CONFIG_PA20 is not defined. Further, when
CONFIG_PA20 is not defined, alternative patching converts the TLB purges to
local purges when PA 2.0 hardware has been detected.

Signed-off-by: John David Anglin <dave.anglin@bell.net>
Signed-off-by: Helge Deller <deller@gmx.de>