FROMLIST: drm/panel: set display info in panel attach
Devicetree systems can set panel orientation via a panel binding, but
there's no way, as is, to propagate this setting to the connector,
where the property need to be added.
To address this, this patch sets orientation, as well as other fixed
values for the panel, in the drm_panel_attach function. These values
are stored from probe in the drm_panel struct.

(am from https://patchwork.kernel.org/patch/11037733/)

BUG=b:129029943,b:133205744
TEST=boot krane

Change-Id: Icf7e3b093e4fc6dfad9be811bdba663cda417e32
Signed-off-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1687290
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>

Conflicts:
  include/drm/drm_panel.h

[rebase54(mka): resolved context deltas
  don't add comments for attach/detach(), they aren't
  callbacks like the other functions]
Signed-off-by: Matthias Kaehlcke <mka@chromium.org>