CHROMIUM: arm64: dts: mt8183: kukui: Add wifi reset pin dt entry
The QCA wifi SDIO driver resets the controller during initialization,
to work around other issues. See commit
5468e784c0 ("FROMGIT: ath10k: sdio: reset chip on power_down()").

This causes a problem on re-init as the card is still in 1.8V mode
and incorrectly sets S18A=1 in CMD5 response. This causes the host
to send CMD11 to try to change voltage to 1.8V, which fails (as
expected), and therefore, the interface gets stuck at 3.3V/50 Mhz,
instead of the desired 1.8V/200Mhz (SDR104).

To workaround this issue in the QCA6174A-3 controller, we toggle
the WL_EN pin of the chip on re-init, which resets the card state,
restores the voltage to 3.3V, and allows the flow above to work
correctly.

BUG=b:131043339
TEST=Boot kukui, cat /sys/kernel/debug/mmc1/ios
     => clock:		200000000 Hz
        actual clock:	191999939 Hz
        timing spec:	6 (sd uhs SDR104)
        signal voltage:	1 (1.80 V)

Change-Id: Ie512f640c2a5f2b37cf8593d647a1e89c3c72505
Signed-off-by: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1657506
Reviewed-by: Claire Chang <tientzu@chromium.org>