platform/chrome: cros_ec_debugfs: Add debugfs entry to retrieve EC uptime
The new debugfs entry 'uptime' is being made available to userspace so that
a userspace daemon can synchronize EC logs with host time.

Signed-off-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
[rework based on Tim's first approach]
Signed-off-by: Enric Balletbo i Serra <enric.balletbo@collabora.com>
Series-to: linux-kernel
Series-cc: chrome-platform, collabora
Series-cc: Tim Wawrzynczak <twawrzynczak@chromium.org>
Commit-notes:
Hi Tim,

Sorry for beating around the bush a bit. I did a small rework of the
latest patch you sent and now is more close to the first versions you
sent. I was spinning around the idea if we should expose the attribute
when the command is supported or not and I ended with the conclusion,
however, I think makes sense for sysfs attributes, but I'd prefer
simplicity in the code for debugfs, at the end, if we try to read a file
for a non supported command we will get an error which is also useful
debug information.

Could you recheck that the patch still works and that I didn't break
anything, please?

Cheers,
~ Enric
END