x86/hpet: Add mode information to 'struct hpet_channel'
The usage of the individual HPET channels is not tracked in a central
place. The information is scattered in different data structures splits the
HPET reservation in the HPET character device split out into several places
which makes the code hard to follow.

Assigning a mode to the channel allows to consolidate the reservation code
and paves the way for further simplifications.

As a first step set the mode of the legacy channels when the HPET is in
legacy mode.

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>