x86/hotplug: Silence APIC and NMI when CPU is dead
In order to support IPI/NMI broadcasting via the shorthand mechanism side
effects of need to mitigated:

 Shorthand IPIs and NMIs hit all CPUs including unplugged CPUs

Neither of those can be handled on unplugged CPUs for obvious reasons.

The IPI mitigation is trivial: Switch the APIC into disabled mode so it
wont deliver any regular vector to the CPU

NMIs are still propagated to the 'dead' CPUs. To mitigate that add a per
cpu variable which marks the CPU as dead on offline and add a check for
that into the NMI handler. If set, just return.

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>