mm: stub out all of swapops.h for !CONFIG_MMU
The whole header file deals with swap entries and PTEs, none of which can
exist for nommu builds.  The current nommu ports have lots of stubs to
allow the inline functions in swapops.h to compile, but as none of this
functionality is actually used there is no point in even providing it.
This way we don't have to provide the stubs for the upcoming RISC-V nommu
port, and can eventually remove it from the existing ports.

Link: http://lkml.kernel.org/r/20190703122359.18200-4-hch@lst.de
Signed-off-by: Christoph Hellwig <hch@lst.de>
Cc: Vladimir Murzin <vladimir.murzin@arm.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Stephen Rothwell <sfr@canb.auug.org.au>